requirements.txt
numpy==1.15.4
matplotlib==3.0.2
Pillow==8.1.2
